Refactor : MainHeader and Header is now combined #43
