+++ src/layout/Layout.js
@@ -5,7 +5,7 @@ import LinearStepper from "../components/common/Stepper";
 import Button from "@mui/material/Button";
 import Box from "@mui/material/Box";
 import StepInfo from "../components/common/StepInfo";
-import { MainHeader } from "./Header";
+import { Header } from "./Header";
 import { Typography } from "@mui/material";
 import { PropTypes } from "prop-types";
 import { activeState, eachStepState, modalState } from "../recoil/commonState";
@@ -34,7 +34,7 @@ export const Layout = () => {
 
   return (
     <StLayout>
-      <MainHeader />
+      <Header main={false} pages={[]} settings={["Logout"]} />
       <ContentsContainer>
         <LinearStepper />
         <StepContainer>
